requirements.txt
requests
urllib3>=1.24.2
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Pillow
beautifulsoup4==4.6.3
tldextract==2.2.0
